adding warn function
warn in some methods about invalid fields.
also add some tests.
