+++ tests/core/Form/Form.spec.ts
@@ -5,6 +5,7 @@ import { Form } from '../../../src/core/Form'
 import generateOptions from '../../../src/helpers/generateOptions'
 import defaultOptionsSource from '../../../src/default-options'
 import { InterceptorManager } from '../../../src/core/InterceptorManager'
+import * as utils from '../../../src/utils'
 
 jest.mock('../../../src/core/Errors')
 jest.mock('../../../src/core/Validator')
@@ -148,84 +149,6 @@ describe('Form.ts', () => {
     )
   })
 
-  it('should call to validate specific field or all the fields', () => {
-    let form = new Form(data) as Form & FormData
-
-    form.validateAll = jest.fn()
-    form.validateField = jest.fn()
-
-    form.validate()
-
-    expect(form.validateAll).toHaveBeenCalledTimes(1)
-    expect(form.validateField).not.toHaveBeenCalled()
-
-    form.validate('first_name')
-
-    expect(form.validateAll).toHaveBeenCalledTimes(1)
-    expect(form.validateField).toHaveBeenCalledTimes(1)
-  })
-
-  it('should validate specific field', () => {
-    let form = new Form({
-      name: {
-        value: 'a',
-        label: 'The Name',
-        rules: [() => true],
-      },
-    }) as Form & { name: string }
-
-    form.$validator.validateField = jest.fn(() => ['error'])
-
-    let isValid = form.validateField('name')
-
-    expect(isValid).toBe(false)
-    expect(form.$errors.unset).toHaveBeenCalledTimes(1)
-    expect(form.$errors.unset).toBeCalledWith('name')
-    expect(form.$errors.push).toHaveBeenCalledTimes(1)
-    expect(form.$errors.push).toBeCalledWith({
-      name: ['error'],
-    })
-    expect(form.$validator.validateField).toBeCalledWith(
-      { label: 'The Name', value: 'a', key: 'name' },
-      form
-    )
-
-    form.$validator.validateField = jest.fn(() => [])
-
-    isValid = form.validateField('name')
-    expect(isValid).toBe(true)
-    expect(form.$errors.push).toHaveBeenCalledTimes(1)
-    expect(form.$errors.unset).toHaveBeenCalledTimes(2)
-  })
-
-  it('should validate all the fields of the form', () => {
-    let form = new Form({
-      name: {
-        value: null,
-        rules: [() => true],
-      },
-      last_name: {
-        value: null,
-        rules: [() => false],
-      },
-    })
-
-    form.validateField = jest
-      .fn()
-      .mockReturnValueOnce(true)
-      .mockReturnValueOnce(true)
-
-    expect(form.validateAll()).toBe(true)
-    expect(form.validateField).toHaveBeenNthCalledWith(1, 'name')
-    expect(form.validateField).toHaveBeenNthCalledWith(2, 'last_name')
-
-    form.validateField = jest
-      .fn()
-      .mockReturnValueOnce(true)
-      .mockReturnValueOnce(false)
-    expect(form.validateAll()).toBe(false)
-  })
-
   it('should change the defaultOptions options of the Form', () => {
     Form.defaults.options.validation.defaultMessage = ({ label, value }) =>
       `${label}: ${value}`
@@ -276,6 +199,16 @@ describe('Form.ts', () => {
     expect(form.isFieldDirty('last_name')).toBe(false)
   })
 
+  it('should warn if field key that passed to isDirtyField is not exists', () => {
+    let warnMock = jest.spyOn(utils, 'warn')
+
+    let form = new Form({ name: null })
+
+    form.isFieldDirty('some_key')
+
+    expect(warnMock).toHaveBeenCalledTimes(1)
+  })
+
   it('should run isFieldDirty (argument passes to "isDirty")', () => {
     let form = new Form(data) as Form & FormData
 
@@ -294,91 +227,6 @@ describe('Form.ts', () => {
     form.last_name = 'somthing else'
 
     expect(form.isDirty()).toBe(true)
-  })
-
-  it('should validate field that was change if the "validation.onFieldChanged" set as true', () => {
-    let form = new Form(data, {
-      validation: {
-        onFieldChanged: true,
-      },
-    })
-
-    form.validateField = jest.fn()
-
-    form.fieldChanged('first_name')
-
-    expect(form.validateField).toHaveBeenCalledTimes(1)
-    expect(form.validateField).toHaveBeenCalledWith('first_name')
-
-    form.assignOptions({
-      validation: {
-        onFieldChanged: false,
-      },
-    })
-
-    form.fieldChanged('first_name')
-
-    expect(form.validateField).toHaveBeenCalledTimes(1)
-  })
-
-  it('should clear field errors after field changed', () => {
-    let form = new Form(data, {
-      validation: {
-        unsetFieldErrorsOnFieldChange: false,
-      },
-    })
-
-    form.fieldChanged('first_name')
-
-    expect(form.$errors.unset).toHaveBeenCalledTimes(0)
-
-    form.assignOptions({
-      validation: {
-        unsetFieldErrorsOnFieldChange: true,
-      },
-    })
-
-    form.fieldChanged('first_name')
-
-    expect(form.$errors.unset).toHaveBeenCalledTimes(1)
-    expect(form.$errors.unset).toHaveBeenCalledWith('first_name')
-  })
-
-  it('should push to touched and set $onFocus when field is on focus', () => {
-    let form = new Form(data)
-
-    form.fieldFocused('first_name')
-
-    expect(form.$onFocus).toBe('first_name')
-    expect(form.$touched.push).toHaveBeenCalledTimes(1)
-    expect(form.$touched.push).toHaveBeenCalledWith('first_name')
-  })
-
-  it('should resetValues $onFocus if the field is on focus and validate the field if "validation.onFieldBlurred" is set', () => {
-    let form = new Form(data, {
-      validation: {
-        onFieldBlurred: false,
-      },
-    })
-
-    form.validateField = jest.fn()
-    form.$onFocus = 'first_name'
-    form.fieldBlurred('first_name')
-
-    expect(form.$onFocus).toBe(null)
-    expect(form.validateField).toHaveBeenCalledTimes(0)
-
-    form.assignOptions({
-      validation: {
-        onFieldBlurred: true,
-      },
-    })
-    form.$onFocus = 'last_name'
-    form.fieldBlurred('first_name')
-
-    expect(form.$onFocus).toBe('last_name')
-    expect(form.validateField).toHaveBeenCalledTimes(1)
-    expect(form.validateField).toHaveBeenCalledWith('first_name')
   })
 
   it('should reset all the form state', () => {
