+++ tests/jest.setup.ts
@@ -1,0 +1,2 @@
+// Mocking the console error - the `warn` function will not log error out
+console.error = jest.fn()
