+++ tests/utils.spec.ts
@@ -1,4 +1,4 @@
-import { isObject } from '../src/utils'
+import { isObject, warn } from '../src/utils'
 
 describe('utils.js', () => {
   it('should determine if value is object', () => {
@@ -7,5 +7,14 @@ describe('utils.js', () => {
     expect(isObject([])).toBe(false)
     expect(isObject(1)).toBe(false)
     expect(isObject('aa')).toBe(false)
+  })
+
+  it('should console.error an error message', () => {
+    warn('random error message')
+
+    expect(console.error).toHaveBeenCalledTimes(1)
+    expect(console.error).toHaveBeenCalledWith(
+      '[Form-wrapper-js warn]: random error message'
+    )
   })
 })
